put material icons in each component that uses them
